requirements.txt
django
djangorestframework
django-countries
pillow